third_party/WebKit/LayoutTests/fast/forms/form-submission-cancelable.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<form method="GET" target="target">
      <input name="query" value="AAA" />
      <input type="submit" id="submitButton"/>
</form>
<form method="GET" target="target1">
      <input name="query1" value="AAA" />
      <input type="submit" id="submitButton1"/>
</form>
<iframe id="target" name="target"></iframe>
<iframe id="target1" name="target1"></iframe>
<script>
description('Test that form submit within onsubmit event handlers are not delayed and sends the form data when invoked');
var count = 2;

function asyncTest(query) {
    if (query === '?query=AAA' || query === '?query1=BBB') {
        if (--count == 0) {
            testPassed('Both query events received, with correct values: AAA and BBB');
            finishJSTest();
        }
    }
}

document.forms[0].onsubmit = function (event) {
    document.forms[0].submit();
    document.forms[0].children.query.value = 'BBB';
    return false;
}

document.forms[1].onsubmit = function (event) {
    document.forms[1].submit();
    document.forms[1].children.query1.value = 'BBB';
    return true;
}

document.getElementById('target').onload = function(event) {
    // This should return back query AAA
    asyncTest(event.target.contentWindow.location.search);
}

document.getElementById('target1').onload = function(event) {
    // This should return back query BBB
    asyncTest(event.target.contentWindow.location.search);
}

window.onload = function() {
    document.getElementById('submitButton').click();
    document.getElementById('submitButton1').click();
}

if (window.testRunner)
    window.jsTestIsAsync = true;
</script>
</body>
</html>
